requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
flask-marshmallow